average rating and such

+++ db/migrate/20120430051607_create_ratings.rb
@@ -3,7 +3,7 @@ class CreateRatings < ActiveRecord::Migration
     create_table :ratings do |t|
       t.integer :user_id
       t.integer :beer_id
-      t.integer :rate
+      t.float   :rate
       t.string  :tasted_on
 
       t.timestamps
